etl/requirements.txt
ipykernel
pandas
requests
dotenv
boto3
pyarrow
sqlalchemy 
snowflake-sqlalchemy
pyyaml
psycopg2